v2: create network of 6 nodes
as we discussed on monday, i've created the network with the following

1. genesis - 1
2. bootnodes - 2
3. initial signers - 2
4. normal nodes - 3

i've combine genesis and 1 bootnode, cuz genesis must be bootnode, if
not theres no way other nodes can talk with genesis

i've also tested the resiliancy of the network. killing 1 bootnode wont
cause any disruption. aka availability and integrity both are
implemented in this iteration of the network design.

some minor changes:
1. remove port exposion in dockerfiles
2. use host network (thats the way this will work on a single host)
3. enode.sh and nodes.sh now have a timeout to recheck if enode://<addr>
   is not valid

Signed-off-by: Abubakar Abubakar Yusif <abubakaryusuf951@gmail.com>